% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NglWidget.R
\name{NglWidget}
\alias{NglWidget}
\title{NglWidget}
\usage{
NglWidget(
  pdbID,
  htmlContainer,
  componentOptions = list(),
  width = NA_integer_,
  height = NA_integer_,
  quiet = TRUE
)
}
\arguments{
\item{pdbID}{character}

\item{htmlContainer}{character, the name of the DOM element, typically a div}

\item{componentOptions}{list of lists, specifying named components of the molecular structure}

\item{width}{integer  initial width of the widget.}

\item{height}{integer initial height of the widget.}
}
\value{
a reference to an htmlwidget.
}
\description{
NglWidget
}
