% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NglWidget.R
\name{shinyOutput,NglWidget-method}
\alias{shinyOutput,NglWidget-method}
\alias{NglWidgetOutput}
\title{Standard shiny ui rendering construct}
\usage{
\S4method{shinyOutput}{NglWidget}(obj)
}
\arguments{
\item{obj}{an NglWidget instance}
}
\value{
a reference to an htmlwidget
}
\description{
Standard shiny ui rendering construct
}
\examples{
\dontrun{
  mainPanel(shinyOutput(nglWidget)), width=10)
}

}
