% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NglWidget.R
\name{renderWidget,NglWidget-method}
\alias{renderWidget,NglWidget-method}
\alias{renderWidget}
\title{More shiny plumbing -  an NglWidget wrapper for htmlwidget standard rendering operation}
\usage{
\S4method{renderWidget}{NglWidget}(obj)
}
\arguments{
\item{expr}{an expression that generates an HTML widget.}

\item{env}{environment in which to evaluate expr.}

\item{quoted}{logical specifies whether expr is quoted ("useuful if you want to save an expression in a variable").}
}
\value{
not sure
}
\description{
More shiny plumbing -  an NglWidget wrapper for htmlwidget standard rendering operation
}
